requirements.txt
aiohttp==3.8.4
aiohttp-apispec==2.2.3
aiosignal==1.3.1
apispec==3.3.2
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
multidict==6.0.4
packaging==23.1
PyYAML==6.0
redis==4.5.5
trafaret==2.1.1
webargs==5.5.3
yarl==1.9.2
